Small chatroom created
This was a small chatroom I created in a Free JavaScript Workshop done by Lighthouse Labs. 

Haha NOTE that most of the design was not written by me. 
The only parts that were done by me is the responses to a question, as well as asking what the date is and what the weather is like.